moving eventful and test modules
