--- timecontrol/logs/events.py
@@ -1,40 +1,0 @@
-import inspect
-from dataclasses import dataclass
-from datetime import datetime, timezone
-import typing
-from result import Result
-
-# TODO : different kinds of !structured! log / dataframes, etc.
-
-
-@dataclass(frozen=True, init=False)
-class CommandCalled:
-    # TODO : more time representation
-    timestamp: typing.Union[int, datetime]
-    bound_args: typing.FrozenSet  # we need a frozenset here to be hashable
-
-    def __init__(self, bound_arguments: inspect.BoundArguments, timestamp: typing.Union[int, datetime] = None):
-        if timestamp is None:
-            timestamp = datetime.now(tz=timezone.utc)
-        object.__setattr__(self, "timestamp", timestamp)
-        # we change bound arguments to a frozenset (to be hashable)
-        object.__setattr__(self, "bound_args", frozenset(bound_arguments.arguments))
-        # This should be all we need. and we don't need mutability as it is already final
-        # Also order doesnt matter as they are already bound to a name (right ?)
-
-
-@dataclass(frozen=True, init=False)
-class CommandReturned:
-    # TODO : more time representation
-    timestamp: typing.Union[int, datetime]
-    # we use result here to keep return "in-band". we dont want "out-of-band" exceptions breaking the control flow...
-    result: Result
-
-    def __init__(self, result: Result, timestamp: typing.Union[int, datetime] = None):
-        if timestamp is None:
-            timestamp = datetime.now(tz=timezone.utc)
-        object.__setattr__(self, "timestamp", timestamp)
-        object.__setattr__(self, "result", result)
-
-
-CommandEvent = typing.Union[CommandCalled, CommandReturned]
